django-api/src/requirements.txt
Django==1.11.24
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
PyJWT==1.5.3
pytz==2017.3
